requirements.txt
streamlit
pandas
requests
networkx
python-dotenv
plotly
